+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## staged
 
+- Add functionality of run_mc_mld_bf to run_mc_mld via multiple dispatch
 - Fixes inconsistency of connections on MATHEMATICAL components, in particular, virtual objects (#280)
 - Add CONTRIBUTING.md
 - Add pull request template